Clear workspaces on reconnect and fetch immediately on return
My thinking here is that if you try connecting to a new URL you still
see the old workspaces until the new connection is successful which
could be confusing (the new URL shows in the text field yet the
workspaces below it belong to something else and the list will no longer
update since the poller was stopped).

Also thought it might make sense to fetch immediately if you come back
to this step since it is not clear that the data there is possibly
stale (especially if you were on a previous/next step for a long time);
I think I would assume a fresh fetch when I interact with the
next/previous buttons but this is just my personal assumption.

Fetching immediately also "fixes" an interesting bug where if you hit
back you are unable to connect to a workspace until the next poll goes
through.  I considered looking more into this but figured it would be
best to block the next step on fresh data anyway.