Read URL and token from config on launch
In case the user has already configured the CLI.  This will only happen
without user input until the first successful connection.

After that the "use existing token" checkbox must be used to read the
token from the config.  It will only be used if it matches the URL.

A secondary effect of the checkbox is that it will not open the browser
window so it can also be used in the case where you already have a token
you want to use but it is not in the CLI config.

Lastly I removed some padding in an attempt to regain the space lost by
the additions.