Move CLI init into background process
This lets you see the initialization progress so if it hangs due to a
timeout for example you are not stuck on the main screen while it
completes.

Next step is probably to make it cancelable.

As a side effect this does change the flow a little; if the token is
invalid/expired we now always show the token dialog again.  I think that
could be a good change anyway though.

Because showing the dialog again results in the browser window opening
repeatedly I made it only open the first time and added a link so they
can open it themselves if they do need a new token.  I also show the
current token so users can validate the bad token.

Also added an error log for when the connection times out.