Use the same port for all executors
Gateway seems to keep the SSH sessions up (as indicated by the Coder
proxy process persisting) and since we have one executor using port 22
and another using port 0 it was spawning two separate processes.

Using the same port causes it to use the existing process.  No idea if
this will really affect much but I guess it does at least help guarantee
that if you were able to list the editors you will be able to connect as
well since it would use the same connection?