Standardize panel width
Before we used an empty cell which created a gap on the right but that
caused the IDE dropdown to change sizes based on its content.

I removed `indent` and added an empty border so we get consistent
padding on both sides although I have no idea if this is the idiomatic
way to do this.