Use ETag for downloading binary
This way we will not need to name the binaries with their versions or
otherwise extract the version from each binary since we can just grab
their hash and set that in the request.  This also means we will not
need to remove old versions as there will only be the one file.