Refactor connect flow somewhat
Mostly trying to reduce the direct uses of the local model in favor of
passing in arguments so they will be easier to test (specifically
askToken is pure now).

Mention use of the local model in the functions where that is the case
and where they use the model I use it for all things (rather than just
the URL but pass in the token for example) since everything we need is
already in the model.

Also rename/suffix some functions to/with connect to match the button.

Remove the progress updates to the indicator since they are
indeterminate.