Tweak open browser logic
Currently if you launch the Coder plugin from Gateway and it
automatically picks up a token from the config or from your last launch
but it is invalid it will open a browser window for a new token which I
think could be jarring.  We should only open a browser window when the
user explicitly wants that by pressing connect for example.