Refactor connect to not use global model
Should make it easier to test.  This meant not calling
askTokenAndConnect because it interacts with the global model.

Initially I was trying to figure out a way to wait progress dialog
background job so you could then retry but I am not sure how to do that
so I went with passing in a callback.

Also launching a new coroutine since otherwise it blocked the current
job and threw a warning, I think it disliked the invokeAndWait of the
token dialog from inside the job.