Surface token source and error
Now it will say whether the token was from the config or was the last
known token and if it fails there will be an error message.  You could
always check the error in the bottom right but this way it is more
obvious why the token dialog has reappeared.

Also if the URL has changed there is no point trying to use the token
we had stored for the previous URL.