Show errors more prominently and show token source (#228)
* Refactor connect to not use global model

Should make it easier to test.  This meant not calling
askTokenAndConnect because it interacts with the global model.

Initially I was trying to figure out a way to wait progress dialog
background job so you could then retry but I am not sure how to do that
so I went with passing in a callback.

Also launching a new coroutine since otherwise it blocked the current
job and threw a warning, I think it disliked the invokeAndWait of the
token dialog from inside the job.

* Surface connection status and errors in table

Instead of just "Nothing to show" it will show the last error or the
status if we are currently trying to connect.

* Surface token source and error

Now it will say whether the token was from the config or was the last
known token and if it fails there will be an error message.  You could
always check the error in the bottom right but this way it is more
obvious why the token dialog has reappeared.

Also if the URL has changed there is no point trying to use the token
we had stored for the previous URL.