Preserve selection when workspace starts/stops (#233)
* Extract randWorkspace to shared class

* Break out workspaces table

I want to write a unit test for the selection logic.

* Preserve selection when workspaces starts or stops

When a workspace starts the workspace row is replaced with a row for the
agent(s) and you have to select it again which is annoying (similarly
for when the workspace stops).  This preserves the selection when the
workspace transitions.