Retry direct connection
This will cover recent connections which connect directly without going
through the whole setup flow.

Pretty much the same logic as for listing editors but we display the
errors in different ways since this all happens in a progress dialog.  I
tried to combine what I could in the retry.

Also the SshException is misleading; it seems to wrap the real error so
unwrap it otherwise it is impossible to tell what is really wrong.  In
particular this is causing us to retry on cancelations.