Store config directory in recent connection
We need this information so we can query the status of recent
connections as they could belong to multiple deployments.

This could end up desyncing if the user manually edits their config file
and changes the global config path in ProxyCommand.

The alternative would be to parse the SSH config to make sure we have
the right config directory but that would mean parsing ProxyCommand to
extract the value of --global-config.

As a fallback for connections that already exist and are not yet stored
with the config directory we could split the host name itself on `--`
since it has the domain in it and join with the default directory but
this could be inaccurate if the default has been changed or if in the
future we change the host name format.