Store name in recent connection
So we can match against the API response.  We could split the hostname
on `--` but there are cases where that will fail (when the name or
domain itself contains -- in specific configurations).

We have to add the config path anyway so this is the best opportunity to
add more information.