Add status and start/stop buttons to recents view (#243)
* Split client out of client service

This way we can create multiple clients on the recent workspaces page
without having to do the whole init thing for each one.

* Store config directory in recent connection

We need this information so we can query the status of recent
connections as they could belong to multiple deployments.

This could end up desyncing if the user manually edits their config file
and changes the global config path in ProxyCommand.

The alternative would be to parse the SSH config to make sure we have
the right config directory but that would mean parsing ProxyCommand to
extract the value of --global-config.

As a fallback for connections that already exist and are not yet stored
with the config directory we could split the host name itself on `--`
since it has the domain in it and join with the default directory but
this could be inaccurate if the default has been changed or if in the
future we change the host name format.

* Store name in recent connection

So we can match against the API response.  We could split the hostname
on `--` but there are cases where that will fail (when the name or
domain itself contains -- in specific configurations).

We have to add the config path anyway so this is the best opportunity to
add more information.

* Standardize some casing

* Ignore null hostnames

I guess this could happen if you manually edit the recents?  In any case
if there is no host name I am not sure there is value in trying to show
the connection and it is making it difficult to check if the workspace
is up.

* Break out toAgentModels

We will need this in the recent connections view as well.

* Split agent status icon and label

So we can show just the icon in the recent connections view.

* Add status and start/stop buttons to recent connections

This relies on some new data being stored with the recent connections so
old connections will be in an unknown state.

* Simplify recent workspace connection constructor