Bypass download if CLI version matches (#245)
* Bypass download if CLI version matches

This is a bit faster than waiting for a 304 but more importantly it
works if the user customizes the download source to be a place that does
not have the same etag support that Coder does.

The test bypasses Windows since we would need to create an executable
there for the test.  We could split out the exec and the parsing
although I am not sure it is worth the noise.  The part most likely to
break between platforms is probably the exec, not the parsing?

* Make invalid JSON test less obscure

* Properly compare semver