Revert platformVersion
If you run the latest plugin on Gateway 231 you get an error about
DeployWithDownload being void when the IDE is not already installed.

Reverting the platform version back to 231 makes the plugin work on both
231 and 232 (the current EAP).

Honestly I do not understand why this is necessary as the API here does
not appear to have changed at all.

Upon doing some research though it appears that the platform version
should match the "since build" version so I downgraded further.  It
seems `minimumWidth` and `maximumWidth` are not available in that
version so our plugin is currently broken there as well.  This is fixed
by using `size` instead.

I think normally the plugin verifier would catch these things and I
tried getting it working by specifying the versions to check but it
seems to be bugged; the only errors it shows appear to be from a failure
to properly resolve dependencies.