Automatically use token on disk if there is one
Rather than asking the user to confirm.  This only happens if we
explicitly want to use an existing token anyway, and "existing" is
defined in the help text as either the token on disk or one the user
already copied, so the extra confirmation to use the token on disk seems
unnecessary.