Remove deployment host
The deployment host is probably not exactly where the IDEs would be
stored if self-serving.  Probably at most they would share a root
domain.  For now remove it, we can add it back or something like it back
if we figure out some usage patterns, although making it configurable is
probably the better bet anyway.