Handle Gateway links (#289)
* Break connection code out into separate class

The connect entrypoint is used both for Gateway
links (jetbrains-gateway://) and internally in our own code at the end
of the wizard and for recent connections, but this will be two separate
sets of parameters (internally we have the CLI set up and pass around
the hostname while the links will only have the workspace ID or name).

So break out the code into a separate class that we can call internally
which will let us dedicate the connect entrypoint to handle the Gateway
links.  There we will set up the CLI and gather the required parameters
before calling the now-broken-out code.

* Break out askToken dialog

So it can be reused in the link flow.

* Add generic function to ask for input

* Add QUERY token source and show deployment host

When you connect via a Gateway link it might not be obvious if the
deployment URL is wrong.

* Tweak client error responses

* Use exception class name when there is no message

Some exceptions have no message, like null pointer exceptions.  Showing
the class name seems more helpful than "no details", since it can save
you a trip to the logs.

* Handle Gateway links

* Fix If-None-Match header name

* Automatically use token on disk if there is one

Rather than asking the user to confirm.  This only happens if we
explicitly want to use an existing token anyway, and "existing" is
defined in the help text as either the token on disk or one the user
already copied, so the extra confirmation to use the token on disk seems
unnecessary.

* Confirm download link if not whitelisted