Break out getting matching agents
I also made a tweak to check that the agent ID is not null since
toAgentModels() will return the workspace without any agent bits set if
there are no agents.

And the wrong error message would show when either the id or name were
missing.  I also flipped them around while fixing this to match the
order above it.