feat: add configuration options to support mtls
adding options to support mtls with the coder server.  This supports
adding PEM certs and keys to the tls requests, and also supports
adding a CA cert to the trust store.  Also allowing for an alternate
hostname that may appear in the certs which is useful for testing or
for non-standard cert usage.