Use delegated properties with recent connections (#337)
From what I understand when you use BaseState you are supposed to use
delegation.  Something must have changed because this code used to work
just as it was, but now it seems without delegation the properties are
ignored and we were getting recent entries that were blank (looked like
this: `<RecentWorkspaceConnection />`).