Refactor settings
You can get things like the data directory without having to ask the CLI
manager now, which I think makes more sense.

It is also a bit easier to test because you do not have to account for
all the fallback directory logic in the CLI manager tests.

This also ensures code will not mutate the settings.