Move template icon download into client
My main goal is to make it easier to make the client not global, and the
template downloader uses the globally configured client.  We could
change it to pass it in but I figure it is not unreasonable for the
client to know how to fetch assets like icons in addition to fetching
responses for API calls.