Refactor wizard steps
The main goal was the ability to use steps separately.  We may only need
the IDE selection step when handling a link, for example.

Each step now takes a specific type of data (if it needs some) then
returns a specific type of data.  The way to tell if a step is ready to
return data is via a callback you pass into the step.

I think more improvements will become apparent once we start using it in
the link handler.