Make steps usable standalone
The previous refactor was close but we needed a bit more refactoring to
make a single step usable on its own, as it will be used in the link
handler.  The main problem was the wrapper that had both steps in it, we
would have to pass in the steps instead which brings us back to that
whole generalization problem where a step has types that lead into the
next step's init.

The solution is to remove the wrapper and wire up the steps wherever
they are needed, whether that is one step or multiple, and the main
change to make that possible is to move the buttons into each step
rather than having them in a wrapper component.

Now you can create any single step and use its output.

This is used in the link flow and the buttons in the step now replace
the default buttons in the dialog wrapper.