[wip] Refactor for Toolbox
I kept our core code intact as much as possible, but this is technically
an entirely different plugin now.

Some code needs to be pulled back in like the link handler, but it is
not split out so importing is tricky.  Going to deal with that later.
Ideally I can find a generic way to break out that code and share
between the two plugins.