Switch from Gson to Moshi
The issue I am having with Gson is that it relies on some unsafe thing
that is not available in all cases.

Since Moshi is by the same folks that wrote the HTTP and REST clients we
are using, it seemed like the best replacement.

Thankfully, for the most part this was a simple search and replace, but
calling toJson and fromJson is a bit more burdensome as you have to call
`adapter` and include the type, and it has no built-in converter for
UUID.  On the flip side, the converters require slightly less code.