Replaced untyped connection parameters map
Provides more type safety, and gives us a place to do some additional
deploy steps in a future commit.

Additionally, make connections with same workspace, project, and IDE
match, since if there are two for the same IDE but one has a download
link and one has a path we should combine them, for example.