Add setup command
To do this, we now handle the download of the IDE so we can figure out
exactly where it is without having to scan through the list of installed
IDEs afterward and match on the IDE product and build number.

This has some other advantages as well since the default deploy
mechanism does not appear to be very robust.

Closes #386