Fix prematurely deleting recent connection
If you view the recent window, it populates with workspaces.  Then if
you create a new workspace, connect to it, then go back to the recent
window it can get removed because it had not yet done another fetch that
includes the new workspace.

To fix, move the deletion from the render phase to the fetch phase.