Consolidate connection errors
This way we can display the same errors on the recent connections and
"Connect to Coder" panels.

Normally an authentication error is only detected when getting me(), but
on the recents page that is just a wasted call, so refactor the errors
to allow detecting a 401 on any request, which is probably better than
relying on only me() anyway.