Fix coroutines when running from File > Remote Development
For whatever reason, cs.launch does not work and you have to pass the
modality state.  Without this state, they do not launch until after the
remote development window is closed.

Using that from a background process fails, however (something about
only being able to access from EDT), so instead of starting/stopping the
poll there, I just always keep it running in the background (and it just
does nothing when there is no client).

For the re-invoking of the token window, I removed the launch entirely.
Could probably use invokeLater() or something but this seems fine?