Rename ideProductCode to ideProduct
For example, ideProductCode.productCode is a bit weird.  This object
contains more than just the product code.

It is still ideProductCode in the recent connections file because there
we do indeed only store the product code, but also we cannot change the
name there anyway to avoid breaking backwards compatibility.