fix: use background ssh channel for link handler
After some investigation into the interaction with the Jetbrains plugin and coder/coder I found a bug where 2 ssh sessions labeled with `jetbrains` for the usage app were being open. We only want to use the main hostname when we want to track user activity, so for any background processes we should be using the background hostname.