Use wildcard SSH config Host entries
This simplifies the written SSH config and avoids the need to make an
API request for every workspace the filter returns. This can remove
minutes from the "Configuring Coder CLI..." step when the user has
access to many workspaces (for example, an admin who wants the option of
connecting to anyone's workspace on a large deployment).

Depends on https://github.com/coder/coder/pull/16088