chore: upgrade Gradle IntelliJ Plugin
This is needed in order to be able to specify "unlimited until builds".
IntelliJ Platform Gradle Plugin 2.x is the build system that supersedes
the Gradle IntelliJ Plugin 1.x, and this is the version that comes with
the ability to provide unlimited until build support.

This is a major overhaul of the DSL, I've tested these changes locally with
GW 2023.3 (the minimum supported version), and I've also compared the plugin.xml
with a previous version to make sure it is generated correctly.