fix:  pass the folder parameter to the IDE&Project dialog (#550)
* chore: use newer snapshot build for testing and building

I've also changed the list of Gateway versions we run the verifier against
to include Gateway 2025.1 while excluding some ancient releases.

Without this commit the PR will fail to build, the declared snapshots are
no longer available in the repository

* fix: pass the folder parameter to the IDE&Project dialog

The value of the `folder` URI parameter is passed to the IDE&Project
dialog when handling URIs. The folder value will be rendered instead
of the default home folder if the value is not blank.

-resolves #466

* chore: next version is 2.20.1