impl: support for unlimited 'Until Build' setting (#549)
* chore: upgrade Gradle to 8.5 the latest version

This upgrade is needed in order to be able to upgrade the gateway/intellij
platform plugins.

* chore: upgrade Gradle IntelliJ Plugin

This is needed in order to be able to specify "unlimited until builds".
IntelliJ Platform Gradle Plugin 2.x is the build system that supersedes
the Gradle IntelliJ Plugin 1.x, and this is the version that comes with
the ability to provide unlimited until build support.

This is a major overhaul of the DSL, I've tested these changes locally with
GW 2023.3 (the minimum supported version), and I've also compared the plugin.xml
with a previous version to make sure it is generated correctly.

* impl: support for unlimited 'Until Build' setting

Every time a new GW version is released we need to change the untilBuild setting.
With latest IntelliJ Platform Plugin we can provide unlimited upper range with no
manual intervention each time a new release is happening.

* chore: simplify build script

Use gateway DSL to provide the platform type and version.
The plugin verification as well as the platform type depends on the versions in the format "2025.x"

* fix: merge from main was botched

* fix: missing junit4 dependency for IntelliJ test framework

We don't have any UI tests but the test task still fails because it misses junit 4 dependencies.
For the unit tests we use kotlin tests library with junit 5 as a backend.

* fix: task was renamed

listProductsReleases was renamed to printProductsReleases in version 2.x

* chore: use latest version 2.6.0 for the platform plugin