impl: verify cli signature  (#562)
* impl: support for downloading and verifying cli signatures

* fix: class cast exception

* impl: embed the pgp public key as a plugin resource

This is the key that validates if the gpg signature was tampered

* chore: fix UTs related to CLI downloading

For one thing some method signature changed, some methods are now suspending functions
that will have to run in a coroutine in the tests. The second big issue is that now
the download function requests user's input via a dialog

* fix: download the correct CLI signature for Windows

The signature for windows CLI follows the format: coder-windows-amd64.exe.asc
Currently it is coded to coder-windows-amd64.asc which means the plugin
always fail to find any signature for windows cli

* chore: next version is 2.22.0

* impl: strict URL validation for the connection screen

This commit rejects any URL that is opaque, not hierarchical, not using http or https
protocol, or it misses the hostname.

* impl: strict URL validation for the URI handling

This commit rejects any URL that is opaque, not hierarchical, not using http or https
protocol, or it misses the hostname.

* fix: transform to url only after we checked the validation result

* chore: update UT expected result