impl: add support for disabling CLI signature verification (#564)
* impl: add new configurable option to disable CLI signature verification

These options are configurable from the Settings page there is no available
shortcut on the main plugin page to discourage the quick disable of CLI verification

* impl: hide configurable fallback if signature verification is disabled

The main plugin screen has a quick shortcut for setting whether the user
wants to fallback on releases.coder.com for signatures if they are not provided by
the main deployment. This checkbox should not be visible if the user wants to disable
signature verification altogether.

* impl: skip signature validation

Signature validation is skipped if the user configured the `disableSignatureVerification` to true.

* chore: update changelog

* chore: next version is 2.22.1

* doc: developer facing documentation for CLI signature verification

* chore: fix UTs