refactor: only login the cli when starting workspaces
The Coder `Recent projects` view polls every 5 seconds for workspaces from
multiple Coder deployment. The polling process also involves a call
to the `cli.login`. The cli is later used to start workspaces if a user clicks on
a project for which the workspace is stopped. The login can be called on demand, only
when a "recent" project is stopped and the user wants to start it. This commit
reduces a lot of overhead associated with spawning cli commands every 5 seconds.