fix: don't create new api keys each time we do workspace polling (#568)
* fix: don't create new api keys each time we do workspace polling

The default behavior for `coder login --token <token>` is to:
- use the provided token temporarily to authenticate the login process
- generate a new session token and stores that for future CLI use
- the original token provided is not stored or reused

The Coder `Recent projects` view polls every 5 seconds for workspaces from
multiple Coder deployment. The polling process also involves a call
to the `cli.login`. The cli is later used start workspaces if user clicks on
a project for which the workspace is stopped. Instead of generating a new token
each time we login we can use the `coder login --use-token-as-session --token <token>` which:
- uses the provided token directly as the session token
- stores the original token for future CLI commands
- no new token is generated

* refactor: only login the cli when starting workspaces

The Coder `Recent projects` view polls every 5 seconds for workspaces from
multiple Coder deployment. The polling process also involves a call
to the `cli.login`. The cli is later used to start workspaces if a user clicks on
a project for which the workspace is stopped. The login can be called on demand, only
when a "recent" project is stopped and the user wants to start it. This commit
reduces a lot of overhead associated with spawning cli commands every 5 seconds.

* chore: next version is 2.22.2