refactor: remove unnecessary SNI manipulation from SSL socket factory
In our codebase we currently have two layers of custom logic:
- one that alters the SNI in the ClientHello (via a custom SSLSocketFactory)
- another that compares an alternate hostname against the SAN entries during client-side certificate verification.

This work was done for one of Coder's clients that wants to do auth via certificates instead of API tokens. After recent
discussions it turns out the SNI manipulation is not needed, we only need to do custom certificate validation.