impl: ability to start a recent workspace connection after ssh wildcard config was changed
Currently, if a user starts a connection with wildcard enabled and then later on it disables
the wildcard config then the recent connections becomes unusable because the hostnames are invalid.
The issue can reproduce the other way around as well (start with wildcard ssh config disabled,
start an IDE and then later on enable wildcard config)

This commit addresses the issue by resolving the hostname on demand when the user wants to
open the remote IDE from the recent connections panel.