impl: add the option to disable ssh wildcard configuration (#584)
* impl: add the option to disable ssh wildcard configuration

It will be used later by the Coder Settings view to allow users
to enable or disable SSH hostname wildcard configuration.

* impl: expose ssh wildcard config in the Settings page

Updated the UI component to allow configuration by the user

* impl: take into account wildcard configuration

when generating the ssh config for Coder Gateway.
Up until now we just checked if the Coder deployment supports
this feature, but now users have to option to continue to
use expanded hostnames in the ssh config.

* fix: force CLI manager to use user settings

CLIManager can be created with default settings (simplifies testing),
among which the ssh wildcard config is enabled. But in reality the
config can be disabled by the user.

* impl: ability to start a recent workspace connection after ssh wildcard config was changed

Currently, if a user starts a connection with wildcard enabled and then later on it disables
the wildcard config then the recent connections becomes unusable because the hostnames are invalid.
The issue can reproduce the other way around as well (start with wildcard ssh config disabled,
start an IDE and then later on enable wildcard config)

This commit addresses the issue by resolving the hostname on demand when the user wants to
open the remote IDE from the recent connections panel.

* chore: update README

* chore: next version is  2.23.0

* fix: don't show twice the connection to the same workspace

Connections started with two different hostnames (because of the ssh wildcard config)
can be rendered twice in the Recent projects panel.
With this commit we ignore the hostname and instead use the workspace name and deployment URL.